+++ docker.sh
@@ -122,6 +122,26 @@ function buildImages {
     buildImage 1.13.1 open php7.3 fpm
     buildImage 1.13.1 open php7.4 apache
     buildImage 1.13.1 open php7.4 fpm
+    buildImage 1.13.2 pro php7.0 apache
+    buildImage 1.13.2 pro php7.0 fpm
+    buildImage 1.13.2 pro php7.1 apache
+    buildImage 1.13.2 pro php7.1 fpm
+    buildImage 1.13.2 pro php7.2 apache
+    buildImage 1.13.2 pro php7.2 fpm
+    buildImage 1.13.2 pro php7.3 apache
+    buildImage 1.13.2 pro php7.3 fpm
+    buildImage 1.13.2 pro php7.4 apache
+    buildImage 1.13.2 pro php7.4 fpm
+    buildImage 1.13.2 open php7.0 apache
+    buildImage 1.13.2 open php7.0 fpm
+    buildImage 1.13.2 open php7.1 apache
+    buildImage 1.13.2 open php7.1 fpm
+    buildImage 1.13.2 open php7.2 apache
+    buildImage 1.13.2 open php7.2 fpm
+    buildImage 1.13.2 open php7.3 apache
+    buildImage 1.13.2 open php7.3 fpm
+    buildImage 1.13.2 open php7.4 apache
+    buildImage 1.13.2 open php7.4 fpm
 }
 
 function pullImages {
@@ -347,6 +367,26 @@ function pushImages {
     pushImage 1.13.1 open php7.3 fpm
     pushImage 1.13.1 open php7.4 apache
     pushImage 1.13.1 open php7.4 fpm
+    pushImage 1.13.2 pro php7.0 apache
+    pushImage 1.13.2 pro php7.0 fpm
+    pushImage 1.13.2 pro php7.1 apache
+    pushImage 1.13.2 pro php7.1 fpm
+    pushImage 1.13.2 pro php7.2 apache
+    pushImage 1.13.2 pro php7.2 fpm
+    pushImage 1.13.2 pro php7.3 apache
+    pushImage 1.13.2 pro php7.3 fpm
+    pushImage 1.13.2 pro php7.4 apache
+    pushImage 1.13.2 pro php7.4 fpm
+    pushImage 1.13.2 open php7.0 apache
+    pushImage 1.13.2 open php7.0 fpm
+    pushImage 1.13.2 open php7.1 apache
+    pushImage 1.13.2 open php7.1 fpm
+    pushImage 1.13.2 open php7.2 apache
+    pushImage 1.13.2 open php7.2 fpm
+    pushImage 1.13.2 open php7.3 apache
+    pushImage 1.13.2 open php7.3 fpm
+    pushImage 1.13.2 open php7.4 apache
+    pushImage 1.13.2 open php7.4 fpm
 }
 
 function pushImage {
@@ -363,6 +403,26 @@ function pushImage {
 }
 
 function printReadme {
+    printSupportedTags 1.13.2 pro php7.4 fpm
+    printSupportedTags 1.13.2 pro php7.4 apache
+    printSupportedTags 1.13.2 pro php7.3 fpm
+    printSupportedTags 1.13.2 pro php7.3 apache
+    printSupportedTags 1.13.2 pro php7.2 fpm
+    printSupportedTags 1.13.2 pro php7.2 apache
+    printSupportedTags 1.13.2 pro php7.1 fpm
+    printSupportedTags 1.13.2 pro php7.1 apache
+    printSupportedTags 1.13.2 pro php7.0 fpm
+    printSupportedTags 1.13.2 pro php7.0 apache
+    printSupportedTags 1.13.2 open php7.4 fpm
+    printSupportedTags 1.13.2 open php7.4 apache
+    printSupportedTags 1.13.2 open php7.3 fpm
+    printSupportedTags 1.13.2 open php7.3 apache
+    printSupportedTags 1.13.2 open php7.2 fpm
+    printSupportedTags 1.13.2 open php7.2 apache
+    printSupportedTags 1.13.2 open php7.1 fpm
+    printSupportedTags 1.13.2 open php7.1 apache
+    printSupportedTags 1.13.2 open php7.0 fpm
+    printSupportedTags 1.13.2 open php7.0 apache
     printSupportedTags 1.13.1 pro php7.4 fpm
     printSupportedTags 1.13.1 pro php7.4 apache
     printSupportedTags 1.13.1 pro php7.3 fpm