+++ docker.sh
@@ -146,11 +146,19 @@ function buildImages {
     buildImage 1.13.2 open php7.3 fpm
     buildImage 1.13.2 open php7.4 apache
     buildImage 1.13.2 open php7.4 fpm
+    buildImage 1.14 pro php7.1 apache
+    buildImage 1.14 pro php7.1 fpm
+    buildImage 1.14 pro php7.2 apache
+    buildImage 1.14 pro php7.2 fpm
+    buildImage 1.14 pro php7.3 apache
+    buildImage 1.14 pro php7.3 fpm
+    buildImage 1.14 pro php7.4 apache
+    buildImage 1.14 pro php7.4 fpm
 }
 
 function pullImages {
-    pullImage php:7.4-rc-fpm-buster
-    pullImage php:7.4-rc-apache-buster
+    pullImage php:7.4-fpm-buster
+    pullImage php:7.4-apache-buster
     pullImage php:7.3-fpm-buster
     pullImage php:7.3-apache-buster
     pullImage php:7.2-fpm-buster
@@ -389,6 +397,14 @@ function pushImages {
     pushImage 1.13.2 open php7.3 fpm
     pushImage 1.13.2 open php7.4 apache
     pushImage 1.13.2 open php7.4 fpm
+    pushImage 1.14 pro php7.1 apache
+    pushImage 1.14 pro php7.1 fpm
+    pushImage 1.14 pro php7.2 apache
+    pushImage 1.14 pro php7.2 fpm
+    pushImage 1.14 pro php7.3 apache
+    pushImage 1.14 pro php7.3 fpm
+    pushImage 1.14 pro php7.4 apache
+    pushImage 1.14 pro php7.4 fpm
 }
 
 function pushImage {
@@ -405,6 +421,14 @@ function pushImage {
 }
 
 function printReadme {
+    printSupportedTags 1.14 pro php7.4 fpm
+    printSupportedTags 1.14 pro php7.4 apache
+    printSupportedTags 1.14 pro php7.3 fpm
+    printSupportedTags 1.14 pro php7.3 apache
+    printSupportedTags 1.14 pro php7.2 fpm
+    printSupportedTags 1.14 pro php7.2 apache
+    printSupportedTags 1.14 pro php7.1 fpm
+    printSupportedTags 1.14 pro php7.1 apache
     printSupportedTags 1.13.2 pro php7.4 fpm
     printSupportedTags 1.13.2 pro php7.4 apache
     printSupportedTags 1.13.2 pro php7.3 fpm