+++ docker.sh
@@ -149,6 +149,14 @@ function buildImages {
     buildImage 1.14.1 pro php7.3 fpm
     buildImage 1.14.1 pro php7.4 apache
     buildImage 1.14.1 pro php7.4 fpm
+    buildImage 1.14.1 open php7.1 apache
+    buildImage 1.14.1 open php7.1 fpm
+    buildImage 1.14.1 open php7.2 apache
+    buildImage 1.14.1 open php7.2 fpm
+    buildImage 1.14.1 open php7.3 apache
+    buildImage 1.14.1 open php7.3 fpm
+    buildImage 1.14.1 open php7.4 apache
+    buildImage 1.14.1 open php7.4 fpm
 }
 
 function pullImages {
@@ -421,6 +429,14 @@ function pushImages {
     pushImage 1.14.1 pro php7.3 fpm
     pushImage 1.14.1 pro php7.4 apache
     pushImage 1.14.1 pro php7.4 fpm
+    pushImage 1.14.1 open php7.1 apache
+    pushImage 1.14.1 open php7.1 fpm
+    pushImage 1.14.1 open php7.2 apache
+    pushImage 1.14.1 open php7.2 fpm
+    pushImage 1.14.1 open php7.3 apache
+    pushImage 1.14.1 open php7.3 fpm
+    pushImage 1.14.1 open php7.4 apache
+    pushImage 1.14.1 open php7.4 fpm
 }
 
 function pushImage {
@@ -437,6 +453,15 @@ function pushImage {
 }
 
 function printReadme {
+    printSupportedTags 1.14.1 open php7.4 fpm
+    printSupportedTags 1.14.1 open php7.4 apache
+    printSupportedTags 1.14.1 open php7.3 fpm
+    printSupportedTags 1.14.1 open php7.3 apache
+    printSupportedTags 1.14.1 open php7.2 fpm
+    printSupportedTags 1.14.1 open php7.2 apache
+    printSupportedTags 1.14.1 open php7.1 fpm
+    printSupportedTags 1.14.1 open php7.1 apache
+
     printSupportedTags 1.14.1 pro php7.4 fpm
     printSupportedTags 1.14.1 pro php7.4 apache
     printSupportedTags 1.14.1 pro php7.3 fpm
@@ -445,14 +470,7 @@ function printReadme {
     printSupportedTags 1.14.1 pro php7.2 apache
     printSupportedTags 1.14.1 pro php7.1 fpm
     printSupportedTags 1.14.1 pro php7.1 apache
-    printSupportedTags 1.14 pro php7.4 fpm
-    printSupportedTags 1.14 pro php7.4 apache
-    printSupportedTags 1.14 pro php7.3 fpm
-    printSupportedTags 1.14 pro php7.3 apache
-    printSupportedTags 1.14 pro php7.2 fpm
-    printSupportedTags 1.14 pro php7.2 apache
-    printSupportedTags 1.14 pro php7.1 fpm
-    printSupportedTags 1.14 pro php7.1 apache
+
     printSupportedTags 1.14 open php7.4 fpm
     printSupportedTags 1.14 open php7.4 apache
     printSupportedTags 1.14 open php7.3 fpm
@@ -461,16 +479,16 @@ function printReadme {
     printSupportedTags 1.14 open php7.2 apache
     printSupportedTags 1.14 open php7.1 fpm
     printSupportedTags 1.14 open php7.1 apache
-    printSupportedTags 1.13.2 pro php7.4 fpm
-    printSupportedTags 1.13.2 pro php7.4 apache
-    printSupportedTags 1.13.2 pro php7.3 fpm
-    printSupportedTags 1.13.2 pro php7.3 apache
-    printSupportedTags 1.13.2 pro php7.2 fpm
-    printSupportedTags 1.13.2 pro php7.2 apache
-    printSupportedTags 1.13.2 pro php7.1 fpm
-    printSupportedTags 1.13.2 pro php7.1 apache
-    printSupportedTags 1.13.2 pro php7.0 fpm
-    printSupportedTags 1.13.2 pro php7.0 apache
+
+    printSupportedTags 1.14 pro php7.4 fpm
+    printSupportedTags 1.14 pro php7.4 apache
+    printSupportedTags 1.14 pro php7.3 fpm
+    printSupportedTags 1.14 pro php7.3 apache
+    printSupportedTags 1.14 pro php7.2 fpm
+    printSupportedTags 1.14 pro php7.2 apache
+    printSupportedTags 1.14 pro php7.1 fpm
+    printSupportedTags 1.14 pro php7.1 apache
+
     printSupportedTags 1.13.2 open php7.4 fpm
     printSupportedTags 1.13.2 open php7.4 apache
     printSupportedTags 1.13.2 open php7.3 fpm
@@ -481,16 +499,18 @@ function printReadme {
     printSupportedTags 1.13.2 open php7.1 apache
     printSupportedTags 1.13.2 open php7.0 fpm
     printSupportedTags 1.13.2 open php7.0 apache
-    printSupportedTags 1.13.1 pro php7.4 fpm
-    printSupportedTags 1.13.1 pro php7.4 apache
-    printSupportedTags 1.13.1 pro php7.3 fpm
-    printSupportedTags 1.13.1 pro php7.3 apache
-    printSupportedTags 1.13.1 pro php7.2 fpm
-    printSupportedTags 1.13.1 pro php7.2 apache
-    printSupportedTags 1.13.1 pro php7.1 fpm
-    printSupportedTags 1.13.1 pro php7.1 apache
-    printSupportedTags 1.13.1 pro php7.0 fpm
-    printSupportedTags 1.13.1 pro php7.0 apache
+
+    printSupportedTags 1.13.2 pro php7.4 fpm
+    printSupportedTags 1.13.2 pro php7.4 apache
+    printSupportedTags 1.13.2 pro php7.3 fpm
+    printSupportedTags 1.13.2 pro php7.3 apache
+    printSupportedTags 1.13.2 pro php7.2 fpm
+    printSupportedTags 1.13.2 pro php7.2 apache
+    printSupportedTags 1.13.2 pro php7.1 fpm
+    printSupportedTags 1.13.2 pro php7.1 apache
+    printSupportedTags 1.13.2 pro php7.0 fpm
+    printSupportedTags 1.13.2 pro php7.0 apache
+
     printSupportedTags 1.13.1 open php7.4 fpm
     printSupportedTags 1.13.1 open php7.4 apache
     printSupportedTags 1.13.1 open php7.3 fpm
@@ -501,16 +521,18 @@ function printReadme {
     printSupportedTags 1.13.1 open php7.1 apache
     printSupportedTags 1.13.1 open php7.0 fpm
     printSupportedTags 1.13.1 open php7.0 apache
-    printSupportedTags 1.13 pro php7.4 fpm
-    printSupportedTags 1.13 pro php7.4 apache
-    printSupportedTags 1.13 pro php7.3 fpm
-    printSupportedTags 1.13 pro php7.3 apache
-    printSupportedTags 1.13 pro php7.2 fpm
-    printSupportedTags 1.13 pro php7.2 apache
-    printSupportedTags 1.13 pro php7.1 fpm
-    printSupportedTags 1.13 pro php7.1 apache
-    printSupportedTags 1.13 pro php7.0 fpm
-    printSupportedTags 1.13 pro php7.0 apache
+
+    printSupportedTags 1.13.1 pro php7.4 fpm
+    printSupportedTags 1.13.1 pro php7.4 apache
+    printSupportedTags 1.13.1 pro php7.3 fpm
+    printSupportedTags 1.13.1 pro php7.3 apache
+    printSupportedTags 1.13.1 pro php7.2 fpm
+    printSupportedTags 1.13.1 pro php7.2 apache
+    printSupportedTags 1.13.1 pro php7.1 fpm
+    printSupportedTags 1.13.1 pro php7.1 apache
+    printSupportedTags 1.13.1 pro php7.0 fpm
+    printSupportedTags 1.13.1 pro php7.0 apache
+
     printSupportedTags 1.13 open php7.4 fpm
     printSupportedTags 1.13 open php7.4 apache
     printSupportedTags 1.13 open php7.3 fpm
@@ -521,6 +543,17 @@ function printReadme {
     printSupportedTags 1.13 open php7.1 apache
     printSupportedTags 1.13 open php7.0 fpm
     printSupportedTags 1.13 open php7.0 apache
+
+    printSupportedTags 1.13 pro php7.4 fpm
+    printSupportedTags 1.13 pro php7.4 apache
+    printSupportedTags 1.13 pro php7.3 fpm
+    printSupportedTags 1.13 pro php7.3 apache
+    printSupportedTags 1.13 pro php7.2 fpm
+    printSupportedTags 1.13 pro php7.2 apache
+    printSupportedTags 1.13 pro php7.1 fpm
+    printSupportedTags 1.13 pro php7.1 apache
+    printSupportedTags 1.13 pro php7.0 fpm
+    printSupportedTags 1.13 pro php7.0 apache
 }
 
 function printSupportedTags {